fix: improve load times and add version in footer
