+++ web/apps/benchmarks-and-reports/next.config.js
@@ -1,5 +1,6 @@
 await import("./src/env.js");
 import withBundleAnalyzer from "@next/bundle-analyzer";
+import packageJson from "./package.json" assert { type: "json" };
 
 /** @type {import("next").NextConfig} */
 const config = {
@@ -10,6 +11,9 @@ const config = {
   transpilePackages: ["@risc0/ui"],
   experimental: {
     caseSensitiveRoutes: true,
+  },
+  publicRuntimeConfig: {
+    version: packageJson.version,
   },
 };
