+++ web/apps/benchmarks-and-reports/src/app/(benchmarks-and-reports)/applications-benchmarks/[[...slug]]/page.tsx
@@ -1,17 +1,14 @@
-import { Link } from "@risc0/ui/link";
 import { Separator } from "@risc0/ui/separator";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@risc0/ui/tabs";
+import { Tabs } from "@risc0/ui/tabs";
 import { truncate } from "@risc0/ui/utils/truncate";
 import type { Metadata } from "next";
 import { redirect } from "next/navigation";
+import { Suspense } from "react";
 import { CopyButton } from "shared/client/components/copy-button";
-import { convertCsvToJson } from "shared/utils/convert-csv-to-json";
 import { replace } from "string-ts";
 import { APPLICATIONS_BENCHMARKS_DESCRIPTION } from "../../_utils/constants";
-import { fetchApplicationsBenchmarks } from "./_actions/fetch-applications-benchmarks";
 import { fetchApplicationsBenchmarksCommitHash } from "./_actions/fetch-applications-benchmarks-commit-hash";
-import { ApplicationsBenchmarksTable } from "./_components/applications-benchmarks-table";
-import { applicationsBenchmarksTableColumns } from "./_components/applications-benchmarks-table-columns";
+import ApplicationsBenchmarksContent from "./_components/applications-benchmarks-content";
 import { FILENAMES_TO_TITLES } from "./_utils/constants";
 
 export const metadata: Metadata = {
@@ -21,9 +18,6 @@ export const metadata: Metadata = {
 
 export default async function ApplicationsBenchmarksPage({ params }) {
   const commitHash = await fetchApplicationsBenchmarksCommitHash();
-  const urls = Object.keys(FILENAMES_TO_TITLES);
-  const dataPromises = urls.map((url) => fetchApplicationsBenchmarks(url));
-  const data = await Promise.all(dataPromises);
 
   if (!params.slug) {
     redirect(`/applications-benchmarks/${replace(Object.keys(FILENAMES_TO_TITLES)[0]!, ".csv", "")}`);
@@ -43,27 +37,9 @@ export default async function ApplicationsBenchmarksPage({ params }) {
       <Separator className="mt-2" />
 
       <Tabs className="mt-6" defaultValue={params.slug?.[0]}>
-        <div className="flex items-center overflow-auto">
-          <TabsList>
-            {Object.keys(FILENAMES_TO_TITLES).map((filename, index) => (
-              <Link key={filename} href={`/applications-benchmarks/${replace(filename, ".csv", "")}`}>
-                <TabsTrigger value={replace(filename, ".csv", "")}>
-                  {Object.values(FILENAMES_TO_TITLES)[index]}
-                </TabsTrigger>
-              </Link>
-            ))}
-          </TabsList>
-        </div>
-
-        {Object.keys(FILENAMES_TO_TITLES).map((filename, index) => (
-          <TabsContent key={filename} value={replace(filename, ".csv", "")} className="mt-4">
-            <ApplicationsBenchmarksTable
-              title={Object.values(FILENAMES_TO_TITLES)[index] ?? ""}
-              columns={applicationsBenchmarksTableColumns}
-              data={convertCsvToJson(data[index]).filter((element) => element.job_name)}
-            />
-          </TabsContent>
-        ))}
+        <Suspense>
+          <ApplicationsBenchmarksContent />
+        </Suspense>
       </Tabs>
     </div>
   );
