+++ web/apps/benchmarks-and-reports/src/app/(benchmarks-and-reports)/crates-validation/page.tsx
@@ -2,15 +2,11 @@ import { Link } from "@risc0/ui/link";
 import { Separator } from "@risc0/ui/separator";
 import { truncate } from "@risc0/ui/utils/truncate";
 import type { Metadata } from "next";
+import { Suspense } from "react";
 import { CopyButton } from "shared/client/components/copy-button";
 import { CRATES_VALIDATION_DESCRIPTION } from "../_utils/constants";
-import { fetchCratesValidationResults } from "./_actions/fetch-crates-validation-results";
 import { findMostRecentHash } from "./_actions/find-most-recent-hash";
-import { CratesIoValidationSummary } from "./_components/crates-io-validation-summary";
-import { CratesIoValidationSummaryHeader } from "./_components/crates-io-validation-summary-header";
-import { CratesIoValidationTable } from "./_components/crates-io-validation-table";
-import { cratesIoValidationTableColumns } from "./_components/crates-io-validation-table-columns";
-import type { CratesIoValidationTableSchema } from "./_components/crates-io-validation-table-schema";
+import CratesIoValidationContent from "./_components/crates-io-validation-content";
 
 export const metadata: Metadata = {
   title: "Crates.io Validation",
@@ -19,7 +15,6 @@ export const metadata: Metadata = {
 
 export default async function CratesIoValidationPage() {
   const mostRecentHash = await findMostRecentHash();
-  const cratesValidationResults: CratesIoValidationTableSchema[] = await fetchCratesValidationResults(mostRecentHash);
 
   return (
     <div className="container max-w-screen-3xl">
@@ -44,9 +39,9 @@ export default async function CratesIoValidationPage() {
       <Separator className="mt-2" />
 
       <div className="mt-6">
-        <CratesIoValidationSummaryHeader data={cratesValidationResults} />
-        <CratesIoValidationSummary data={cratesValidationResults} />
-        <CratesIoValidationTable data={cratesValidationResults} columns={cratesIoValidationTableColumns} />
+        <Suspense>
+          <CratesIoValidationContent mostRecentHash={mostRecentHash} />
+        </Suspense>
       </div>
     </div>
   );
