+++ web/package.json
@@ -8,22 +8,22 @@
     "typecheck": "turbo typecheck"
   },
   "dependencies": {
-    "@next/bundle-analyzer": "14.1.4",
+    "@next/bundle-analyzer": "14.2.1",
     "@poppinss/intl-formatter": "3.0.2",
-    "@risc0/ui": "0.0.48",
+    "@risc0/ui": "0.0.49",
     "@t3-oss/env-nextjs": "0.9.2",
     "@tanstack/match-sorter-utils": "8.15.1",
-    "@tanstack/react-table": "8.15.3",
+    "@tanstack/react-table": "8.16.0",
     "@types/chart.js": "2.9.41",
     "@types/jest": "29.5.12",
     "@types/node": "20.12.7",
-    "@types/react": "18.2.75",
-    "@types/react-dom": "18.2.24",
+    "@types/react": "18.2.78",
+    "@types/react-dom": "18.2.25",
     "@vercel/analytics": "1.2.2",
     "@vercel/edge-config": "1.1.0",
     "chart.js": "2.9.4",
     "deepmerge": "4.3.1",
-    "lucide-react": "0.367.0",
+    "lucide-react": "0.368.0",
     "next": "canary",
     "next-themes": "0.3.0",
     "nextjs13-progress": "1.2.5",
